separate quote storing implementation into separate module
