+++ crates/orderbook/src/database.rs
@@ -2,6 +2,7 @@ pub mod events;
 pub mod fees;
 pub mod instrumented;
 pub mod orders;
+pub mod quotes;
 pub mod trades;
 
 use anyhow::Result;
